* centralize models variables  * rename some functions  * collect alter_views conditionally and recreate in the end  * extract get_view_field_state_write_fields  * debug main functions with make test-file
